+++ server/server.go
@@ -26,6 +26,7 @@ import (
 // Options used for starting Simple IoT
 type Options struct {
 	StoreFile         string
+	ResetStore        bool
 	DataDir           string
 	HTTPPort          string
 	DebugHTTP         bool
@@ -76,7 +77,11 @@ func NewServer(o Options) (*Server, *nats.Conn, error) {
 		nats.ReconnectWait(time.Millisecond*250),
 		nats.ErrorHandler(func(_ *nats.Conn,
 			sub *nats.Subscription, err error) {
-			log.Printf("NATS server client error, sub: %v, err: %s\n", sub.Subject, err)
+			var subject string
+			if sub != nil {
+				subject = sub.Subject
+			}
+			log.Printf("NATS server client error, sub: %v, err: %s\n", subject, err)
 		}),
 		nats.ReconnectHandler(func(_ *nats.Conn) {
 			log.Println("Nats server client reconnect")
@@ -183,6 +188,12 @@ func (s *Server) Start() error {
 
 	siotStore, err := store.NewStore(storeParams)
 
+	if o.ResetStore {
+		if err := siotStore.Reset(); err != nil {
+			log.Fatal("Error resetting store:", err)
+		}
+	}
+
 	if err != nil {
 		log.Fatal("Error creating store: ", err)
 	}
@@ -224,7 +235,7 @@ func (s *Server) Start() error {
 			return nil
 		}
 
-		rootNode, err := client.GetNode(s.nc, "root", "")
+		rootNode, err := client.GetNodes(s.nc, "root", "all", "", false)
 
 		if err != nil {
 			logLS("LS: Exited: store metrics")